log console, eventbus

+++ app/MemoryFileSystem.py
@@ -1,25 +1,31 @@
 import os
 import shutil
 import string
-import sys
 import time
 from ctypes import windll
 
+import numpy as np
 from fs.osfs import OSFS
+
+from app import EventBus
+from app.SGMEvents.MFSEvents import *
 
 
 class MemoryFileSystem:
     hidden_tag_file = '.tag-ram'
-    possible_drives = ['R', 'S', 'T', 'V']
-    ram_drive: str | None = None
+    possible_drives = ['R', 'S', 'T', 'V']  # https://en.wikipedia.org/wiki/Drive_letter_assignment#Common_assignments
+    ram_drive: str = None
 
-    def __init__(self, save_path, backup_path):
+    def __init__(self, save_path, backup_path, event_bus: EventBus):
+        self._event_bus = event_bus
         self.save_path = save_path
         self.backup_path = backup_path
+
+    def run(self):
         self.create_or_just_get()
         time.sleep(0.3)
-        if self.delete_save_folder():
-            self.create_symlink()
+        if self._delete_save_folder():
+            self._create_symlink()
 
     # noinspection SpellCheckingInspection
     # these drive can be possible ram drives
@@ -43,30 +49,36 @@ class MemoryFileSystem:
         else:
             raise Exception('no free drive')
 
-    def create_ram_drive(self, letter) -> bool:
-        success = os.system(f'imdisk -a -s 512M -m "{letter}:" -p "/fs:NTFS /V:SaveGameManager /Q /y"') == 0
+    def create_ram_drive(self, letter, size: int = 512) -> bool:
+        size = np.clip(size, 4, 8192)
+        success = os.system(f'imdisk -a -s {size}M -m "{letter}:" -p "/fs:NTFS /V:SaveGameManager /Q /y"') == 0
         if success and os.path.ismount(f'{letter}:'):
             OSFS(f'{letter}:\\').create(self.hidden_tag_file)
             os.system(f'attrib +H ' + letter + ':\\' + self.hidden_tag_file)
             self.ram_drive = letter
+            self._event_bus.emit(MFSDriveCreated)
             return True
         return False
 
     def destroy_ram_drive(self) -> bool:
         if self.ram_drive:
-            return os.system(f'imdisk -D -m "{self.ram_drive}:"') == 0
+            ret = os.system(f'imdisk -D -m "{self.ram_drive}:"') == 0
+            if ret:
+                self._event_bus.emit(MFSDriveDestroyed)
+            return ret
         return True
 
-    def create_symlink(self) -> bool:
+    def _create_symlink(self) -> bool:
         try:
             os.symlink(self.ram_drive + ':\\', self.save_path)
             if os.readlink(self.save_path):
+                self._event_bus.emit('mfs.symlink.created')
                 return True
         finally:
             pass
         return False
 
-    def delete_save_folder(self) -> bool:
+    def _delete_save_folder(self) -> bool:
         if os.path.isdir(self.save_path):
             if not os.path.isdir(self.backup_path):
                 os.mkdir(self.backup_path)
@@ -79,22 +91,23 @@ class MemoryFileSystem:
             return True
         return False
 
-    def restore_save_empty(self):
-        try:
+    def _restore_save_empty(self) -> bool:
+        if os.path.islink(self.save_path):
             os.rmdir(self.save_path)
             os.mkdir(self.save_path)
-        finally:
-            pass
+            self._event_bus.emit('mfs.symlink.removed')
+            return True
+        return False
 
     def __del__(self):
         self.destroy_ram_drive()
-        self.restore_save_empty()
+        self._restore_save_empty()
 
 
 if __name__ == '__main__':
     p = 'E:\\bla\\save'
     b = 'E:\\bla\\save-backup'
-    mfs = MemoryFileSystem(p, b)
+    mfs = MemoryFileSystem(p, b, None)
     time.sleep(3)
     del mfs
